Add print statements for every function
Will be useful for debugging tomorrow. Data should
all be there, just need to make sure calculations
go smoothly. Then I can work on interactivity but for now, putting the
user on rails with a detailed list of instructions is probably the best
option. README would be a good place to put that.